Exercise5: Change structure of task4
